ddl/placement: cleanup pingcap/check dep for rule_test.go
